Fix arguments
Arguments were being added to Command by name.
Arguments were being added to Parameter by index.
Index was thus not an int, but a name.
A non-numerical string converted to int becomes 0.
Thus, all arguments in Parameter were at order 0.
Thus, while there were possibly a lot of arguments in Command, there was only 1 in Parameter.
Why didn't the test pick this up? Because the test case didn't represent the actual situation.
As soon as I added names as keys to the test, it broke.

Solution: don't key by name in Command.
It wasn't used anyway, and the typehint for getArguments was `Argument[]` which implies numerical keys.
Not adding a changelog, because the bug was introduced earlier in PR #27 and hadn't reached master yet.